refactor(analysis): extract derived file metrics owner
